+++ src/plugins/lab.ts
@@ -1,5 +1,5 @@
 import { rgb2lab } from "@models/rgb";
-import { parseLABColor, roundLAB } from "@models/lab";
+import { parseLABColor, roundLAB, rgb2labString } from "@models/lab";
 import type { ColorLAB, Plugin } from "../types";
 
 declare module "../blossom" {
@@ -8,6 +8,13 @@ declare module "../blossom" {
 		 * Converts a color to CIE LAB color space.
 		 */
 		lab: ColorLAB;
+
+		/**
+		 * Returns a CIE LAB color string.
+		 * 
+		 * https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/lab()
+		 */
+		toStringLAB: string;
 	}
 }
 
@@ -20,6 +27,12 @@ export const pluginLAB: Plugin = (BaseClass, parsers) => {
 	Object.defineProperty(BaseClass.prototype, "lab", {
 		get: function lab() {
 			return roundLAB(rgb2lab(this.color));
+		}
+	});
+
+	Object.defineProperty(BaseClass.prototype, "toStringLAB", {
+		get: function toStringLAB() {
+			return rgb2labString(this.color);
 		}
 	});